Get php unit running on php 8.1 (#3)
* Added \ReturnTypeWillChange Attribute

- Added \ReturnTypeWillChange Attribute
- Added \ReturnTypeWillChange polyfill
- Fixed tests

* Required jworman/polyfill-php81

Required jworman/polyfill-php81

* More ReturnTypeWillChange

* All tests but Amf are working

* Fixed linting errors

* Phpunit tests passing on php 8.1

* Change name

+++ packages/zend-locale/library/Zend/Locale/Math/PhpMath.php
@@ -200,8 +200,8 @@ class Zend_Locale_Math_PhpMath extends Zend_Locale_Math
         if ((int)$op2 == 0) {
             return NULL;
         }
-        $result = $op1 % $op2;
-        if (is_nan($result)  or  (($op1 - $result) % $op2 != 0)) {
+        $result = (int) $op1 % (int) $op2;
+        if (is_nan($result)  or  ((int) ($op1 - $result) % (int) $op2 != 0)) {
             // require_once 'Zend/Locale/Math/Exception.php';
             throw new Zend_Locale_Math_Exception("modulus calculation error: $op1 % $op2 != $result", $op1, $op2, $result);
         }
